+++ requestsink-jms/src/test/java/no/mnemonic/messaging/requestsink/jms/JMSRequestSinkJavaSerializationProxyTest.java
@@ -21,19 +21,22 @@ public class JMSRequestSinkJavaSerializationProxyTest extends AbstractJMSRequest
 
     //set up a real JMS connection to a vm-local activemq
     queueName = "dynamicQueues/" + generateCookie(10);
+    topicName = "dynamicTopics/" + generateCookie(10);
 
     //set up request sink pointing at a vm-local topic
     requestSink = addConnection(JMSRequestSink.builder())
             .setProtocolVersion(ProtocolVersion.V3)
             .setSerializer(new DefaultJavaMessageSerializer())
-            .setDestinationName(queueName)
+            .setQueueName(queueName)
+            .setTopicName(topicName)
             .build();
 
     //set up request proxy listening to the vm-local topic, and pointing to mock endpoint
     requestProxy = addConnection(JMSRequestProxy.builder())
             .addSerializer(new DefaultJavaMessageSerializer())
             .setMaxMessageSize(1000)
-            .setDestinationName(queueName)
+            .setQueueName(queueName)
+            .setTopicName(topicName)
             .setRequestSink(endpoint)
             .build();